+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### New
 
 - \#5 Cache databases for disconnected (offline) usage
+- \#61 Simple Automatic Backup (Remote and/or Local)
 
 ### Fixed